# backend/docs/openapi/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/v1/change-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "req",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/authcontroller.ChangePassword.Req"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/confirm-forgot-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.ConfirmForgotPasswordIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/confirm-sign-up": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.ConfirmSignUpIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/forgot-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.ForgotPasswordIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/healthy": {
            "get": {
                "tags": [
                    "general"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/er.APIError"
                        }
                    }
                }
            }
        },
        "/v1/invite-user": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.LoginIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logic.LoginOut"
                        }
                    }
                }
            }
        },
        "/v1/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.LoginIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logic.LoginOut"
                        }
                    }
                }
            }
        },
        "/v1/ping": {
            "get": {
                "tags": [
                    "general"
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/refresh-token": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "req",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.RefreshTokenIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/logic.RefreshTokenOut"
                        }
                    }
                }
            }
        },
        "/v1/resend-confirmation-code": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.ResendConfirmationCodeIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/set-new-password": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "request body",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.SetNewPasswordIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/v1/sign-up": {
            "post": {
                "tags": [
                    "auth"
                ],
                "parameters": [
                    {
                        "description": "req",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/logic.SignUpIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "authcontroller.ChangePassword.Req": {
            "type": "object",
            "required": [
                "newPass",
                "oldPass"
            ],
            "properties": {
                "newPass": {
                    "type": "string"
                },
                "oldPass": {
                    "type": "string"
                }
            }
        },
        "logic.ConfirmForgotPasswordIn": {
            "type": "object",
            "required": [
                "code",
                "email",
                "newPass"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "newPass": {
                    "type": "string"
                }
            }
        },
        "logic.ConfirmSignUpIn": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "logic.ForgotPasswordIn": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "logic.LoginIn": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "logic.LoginOut": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "idToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "logic.RefreshTokenIn": {
            "type": "object",
            "required": [
                "email",
                "refreshToken"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "logic.RefreshTokenOut": {
            "type": "object",
            "properties": {
                "accessToken": {
                    "type": "string"
                },
                "idToken": {
                    "type": "string"
                },
                "refreshToken": {
                    "type": "string"
                }
            }
        },
        "logic.ResendConfirmationCodeIn": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "logic.SetNewPasswordIn": {
            "type": "object",
            "required": [
                "email",
                "newPass",
                "oldPass"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "newPass": {
                    "type": "string"
                },
                "oldPass": {
                    "type": "string"
                }
            }
        },
        "logic.SignUpIn": {
            "type": "object",
            "required": [
                "email",
                "nickName",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickName": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "er.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "msg": {
                    "type": "string"
                },
                "traceID": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}